isolate model from args, def, train..
