+++ vae_pytorch.py
@@ -5,11 +5,12 @@ from __future__ import print_function
 import argparse
 import os
 import torch
-import torch.utils.data
-from torch import nn, optim
-from torch.nn import functional
+from torch import optim
+
 from torchvision import datasets, transforms
 from torchvision.utils import save_image
+
+from vae_model import VAE_Model
 
 
 # get the arguments, if not on command line, the arguments are the default
@@ -49,43 +50,8 @@ test_loader = torch.utils.data.DataLoader(
     batch_size=args.batch_size, shuffle=True, **kwargs)
 
 
-# definition of beta, should be 0 < beta < 1
-# (this needs to be changed though, could be + pretty if passed as arg)
-
-beta = 0.5
-
-
-class VAE(nn.Module):
-    def __init__(self):
-        super(VAE, self).__init__()
-
-        self.fc1 = nn.Linear(784, 400)
-        self.fc21 = nn.Linear(400, 20)
-        self.fc22 = nn.Linear(400, 20)
-        self.fc3 = nn.Linear(20, 400)
-        self.fc4 = nn.Linear(400, 784)
-
-    def encode(self, x):
-        h1 = functional.relu(self.fc1(x))
-        return self.fc21(h1), self.fc22(h1)
-
-    def reparameterize(self, mu, logvar):
-        std = torch.exp(0.5*logvar)
-        eps = torch.randn_like(std)
-        return mu + beta*eps*std
-
-    def decode(self, z):
-        h3 = functional.relu(self.fc3(z))
-        return torch.sigmoid(self.fc4(h3))
-
-    def forward(self, x):
-        mu, logvar = self.encode(x.view(-1, 784))
-        z = self.reparameterize(mu, logvar)
-        return self.decode(z), mu, logvar
-
-
 # send the model to device
-model = VAE().to(device)
+model = VAE_Model().to(device)
 
 # set the optimizer
 optimizer = optim.Adam(model.parameters(), lr=1e-3)
@@ -93,7 +59,7 @@ optimizer = optim.Adam(model.parameters(), lr=1e-3)
 
 # Reconstruction + KL divergence losses summed over all elements and batch
 def loss_function(recon_x, x, mu, logvar):
-    BCE = functional.binary_cross_entropy(recon_x, x.view(-1, 784), reduction='sum')
+    BCE = torch.nn.functional.binary_cross_entropy(recon_x, x.view(-1, 784), reduction='sum')
     KLD = -0.5 * torch.sum(1 + logvar - mu.pow(2) - logvar.exp())
     return BCE + KLD
